Use standard resource summary on project overview tab
