--- frontend/public/components/_group.scss
@@ -1,26 +1,0 @@
-.group {
-  margin-bottom: ($grid-gutter-width / 2);
-}
-
-.group__body {
-  border: 1px solid $color-grey-background-border;
-  border-top: none;
-}
-
-.group__namespace-details {
-  padding-top: 12px;
-}
-
-.group__title {
-  background-color: $color-pf-black-150;
-  border: 1px solid $color-grey-background-border;
-  display: flex;
-  justify-content: space-between;
-  padding: 10px 15px;
-  @media(max-width: $screen-xs) {
-    flex-direction: column;
-  }
-  h2 {
-    margin: 0;
-  }
-}
