+++ frontend/public/components/dashboard/project-dashboard/project-dashboard.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import * as _ from 'lodash-es';
 
 import Dashboard from '@console/shared/src/components/dashboard/Dashboard';
 import DashboardGrid from '@console/shared/src/components/dashboard/DashboardGrid';
@@ -11,7 +12,6 @@ import { ActivityCard } from './activity-card';
 import { ProjectDashboardContext } from './project-dashboard-context';
 import { LauncherCard } from './launcher-card';
 import { connect } from 'react-redux';
-import { getNamespaceDashboardConsoleLinks } from '../../overview/namespace-overview';
 import { RootState } from '../../../redux';
 
 const mainCards = [{ Card: StatusCard }, { Card: UtilizationCard }];
@@ -21,6 +21,19 @@ const rightCards = [{ Card: ActivityCard }];
 const mapStateToProps = ({ UI }: RootState): ProjectDashboardReduxProps => ({
   consoleLinks: UI.get('consoleLinks'),
 });
+
+const getNamespaceDashboardConsoleLinks = (
+  ns: K8sResourceKind,
+  consoleLinks: K8sResourceKind[],
+): K8sResourceKind[] => {
+  return _.filter(consoleLinks, (link: K8sResourceKind) => {
+    if (link.spec.location !== 'NamespaceDashboard') {
+      return false;
+    }
+    const namespaces: string[] = _.get(link, ['spec', 'namespaceDashboard', 'namespaces']);
+    return _.isEmpty(namespaces) || _.includes(namespaces, ns.metadata.name);
+  });
+};
 
 const ProjectDashboard_: React.FC<ProjectDashboardReduxProps & ProjectDashboardProps> = ({
   obj,
