+++ frontend/public/style.scss
@@ -50,7 +50,6 @@
 @import "components/image-stream";
 @import "components/about-modal";
 @import "components/build-pipeline";
-@import "components/group";
 @import "components/configure-ports";
 @import "components/detail-table";
 @import "components/dropdown";
